Traits~!
Makes it so that you can take up to 3 positive traits, and an unlimited number of negative traits.

Trait points are no longer considered. Instead we can rely on excluding traits we don't want taken together. At the moment there have been no changes made relating to trait exclusions, but there may be some balancing to be done in the future. The main one I might have tried to account for before was haste+hp up+resistance up, but with haste gone, I didn't see it as necessary for now. We'll see how things go!