Less camping
Makes it so that, when playing as very limited (command/adjacent) jobs, on the following shift, one will need to wait 10 minutes before being able to play the same job.

This should help mitigate situations where someone wants to play a limited job, but it's always held by the same individual.

While it's not against the rules or even especially discouraged to play the same job multiple shifts in a row, it can be discouraging to others who might otherwise like to play. Therefore I thought it might be good to allow a short period where someone else can take a turn if they want to.

This is far from a perfect solution. I'm not especially happy about making it a waiting period, but that seemed the best balance, since we do not have a long pregame lobby, so relying on something like a weighted chance on readying up wouldn't be very fit for purpose here.

I did make it so that it will not restrict you on the following shift if you joined within the last 30 minutes. So if you join on the very tail end of a shift, you won't have to worry about waiting at the start of the next one, though, this element has not been widely tested locally. I did confirm it works, but there might be some conditions where it will need some tweaking.