Tailsock support for suits
Only suits that need a tailsock will have the tailsockcolor var with anything in it. This saves on having to do additional checks.

Tails aren't all unified, so I'll have to go through and squash 'em down to be a singular tail sprite to be colored. Since we don't need markings to show up.

Deleted empty bio_vr file

preparing taur bodies that I CBA'd to customize 1500 suits and dresses for. they'll be clipped like usual but now with a full body covering sock to provide 'insulation' in space suits, etc.

Organized the taur dmis a bit, for easier readability. This will continue as I fill in missing sprites