Stops 'It Won't Budge!' messages from gunlockers
I don't know why it was calling relaymove so aggressively

It didn't do this at all in my PR testing. At all.

but all of a sudden putting a gun in the rack would just break closet code entirely. So I've just hijacked those procs so the lockers are entirely self-contained. It stopped the messages! Yaaaay

No runtimes the entire debug, btw. It's been real helpful