Bone related
An idea for a problem - Requires discussion with staff

When a bone is generated via digestion, it registers the predator to a variable, and begins listening to the signal for the predator's movement proc, and destroy proc.

When the predator moves, if the bone is left on a turf, and if the predator is more than 7 turfs away from the bone on x or y, then the bone will delete itself. The bone will not delete itself unless it is on a turf, so you can give your bone to your friend and it will be safe (as long as they don't drop it on the floor while you are moving)

Additionally, makes bones delete themselves when you are deleted. This way if you cryo, all of the bones registered to you, no matter where they are, will be removed. There isn't really a difference between cryoing and other methods of being deleted, so, things like digestion and gibbing will also result in the bones being deleted, but there's not a whole lot I can do about that.

Right now the bones just instantly disappear. Maybe we can make them display a message and fade out or turn into ash instead if just instantly disappearing.

The idea behind this is to have mechanics better represent the intended use case for these mechanics, and hopefully resolve a long standing issue with the digestion remains mechanic. Some people like it, some people can't stand it. Rather than making convoluted and difficult to define and follow rules about how the mechanic can be used, I am seeking a mechanical solution to the problem.

The intention for these bones is, that you can have them around you, you can hide them away somewhere, you can give them to people. You should not leave them out in the open somewhere for others to find.