Enhances admin spawn atom verb
Shortens the paths listed in the spawn atom verb. This was apparently already here but no-one updated the verb to utilize it. So I've fixed that

if you add a : (number) you will spawn that many items at once. Limit 100.

I do not take any responsibility for incidents this may result in.