Revises Hypospray mkii performance
After obtaining feedback over a lengthy period of time, this PR is set to ensure that syringes have actual competition in medicine application.  Delay was greatly reduced to use the hyposprays, The kits can now also hold 14 vials at once, on par with a regular medkit holding a bunch of bottles.

It also greatly improves availability of hypospray mkiis and kit boxes by adding a lot of empty ones to the nano med, hypovials can be produced from the autolathe.

Adds the CMO's version of the mkii into their locker, as well as bundles up the mki version into a box with some extra vials for those who prefer the older over the new one.

Also removes the extra pair of white shoes because clothing vendors exist