Initial
Rebalances Osteodaxon

Now osteo requires 4 units to be metabolized before it will repair bones when injected.

8 Units are needed for use in pill form.

While processing, osteo will also cause halloss and give unique pain messages.  The pain is less when it's in pill form, to balance the higher unit cost.

The intent is to require anesthetics or oxycodone when running an efficient 4 unit injection.