Enables Prometheans & Proteans to commit flight vore
Does what it says on the tin.

If you wanted to do something like, say, a promethean or protean that has wings that can fly, you couldn't do flight vore (or use it for forced slip-vore) because prometheans/proteans couldn't select this trait, only custom species.

I mention this because I was going to make a promethean that can do flight vore and realized I was unable to do such!